# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the EMC Load Balancer Service.",
        "title": "EMC Load Balancer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all orders in the system (requires manage_orders permission)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new order from cart items. Orders are split by shop automatically.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order details with items and optional coupon",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all orders for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List my orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single order by its ID (owner or admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an order's status (requires manage_orders permission)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payment/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a checkout session with SePay Payment Gateway and returns the checkout URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Initialize SePay checkout",
                "parameters": [
                    {
                        "description": "Checkout details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CheckoutInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payment/ipn": {
            "post": {
                "description": "Receives Instant Payment Notification from SePay when payment status changes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "SePay IPN Webhook",
                "parameters": [
                    {
                        "description": "IPN payload from SePay",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.SePayIPNRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payment/test-checkout": {
            "get": {
                "description": "Renders an auto-submitting HTML form that redirects to SePay for testing",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Test checkout link",
                "responses": {
                    "200": {
                        "description": "HTML redirect form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emc_lb_src_pkg_entities.CheckoutInitRequest": {
            "type": "object",
            "required": [
                "order_amount",
                "order_description",
                "order_invoice_number"
            ],
            "properties": {
                "cancel_url": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "error_url": {
                    "type": "string"
                },
                "order_amount": {
                    "type": "number"
                },
                "order_description": {
                    "type": "string"
                },
                "order_invoice_number": {
                    "type": "string"
                },
                "payment_method": {
                    "description": "BANK_TRANSFER, CARD, etc.",
                    "type": "string"
                },
                "success_url": {
                    "type": "string"
                }
            }
        },
...
        "emc_lb_src_pkg_entities.CreateOrderRequest": {
            "type": "object",
            "required": [
                "contact_phone",
                "items",
                "payment_method",
                "shipping_address"
            ],
            "properties": {
                "contact_phone": {
                    "type": "string"
                },
                "coupon_code": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/emc_lb_src_pkg_entities.OrderItem"
                    }
                },
                "payment_method": {
                    "type": "string"
                },
                "shipping_address": {
                    "type": "string"
                }
            }
        },
...
        "emc_lb_src_pkg_entities.OrderItem": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "price": {
                    "description": "Captured at the time of order",
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "emc_lb_src_pkg_entities.SePayIPNOrder": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "order_amount": {
                    "type": "string"
                },
                "order_currency": {
                    "type": "string"
                },
                "order_description": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "order_invoice_number": {
                    "type": "string"
                },
                "order_status": {
                    "description": "CAPTURED",
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "emc_lb_src_pkg_entities.SePayIPNRequest": {
            "type": "object",
            "properties": {
                "notification_type": {
                    "description": "ORDER_PAID",
                    "type": "string"
                },
                "order": {
                    "$ref": "#/definitions/emc_lb_src_pkg_entities.SePayIPNOrder"
                },
                "timestamp": {
                    "type": "integer"
                },
                "transaction": {
                    "$ref": "#/definitions/emc_lb_src_pkg_entities.SePayIPNTx"
                }
            }
        },
        "emc_lb_src_pkg_entities.SePayIPNTx": {
            "type": "object",
            "properties": {
                "authentication_status": {
                    "type": "string"
                },
                "card_brand": {
                    "type": "string"
                },
                "card_expiry": {
                    "type": "string"
                },
                "card_holder_name": {
                    "type": "string"
                },
                "card_number": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "transaction_amount": {
                    "type": "string"
                },
                "transaction_currency": {
                    "type": "string"
                },
                "transaction_date": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_status": {
                    "description": "APPROVED",
                    "type": "string"
                },
                "transaction_type": {
                    "type": "string"
                }
            }
        },
...
        "emc_lb_src_pkg_res.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/emc_lb_src_pkg_res.ErrorCode"
                },
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/emc_lb_src_pkg_res.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {},
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "emc_lb_src_pkg_res.ErrorCode": {
            "type": "string",
            "enum": [
                "COMMON_BAD_REQUEST",
                "COMMON_NOT_FOUND",
                "COMMON_CONFLICT",
                "COMMON_INTERNAL_SERVER_ERROR",
                "COMMON_UNAUTHORIZED",
                "COMMON_FORBIDDEN",
                "COMMON_TOO_MANY_REQUESTS",
                "COMMON_INVALID_IDENTITY_TYPE"
            ],
            "x-enum-varnames": [
                "ErrCodeBadRequest",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeInternal",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeTooManyRequests",
                "ErrCodeInvalidIdentityType"
            ]
        },
        "emc_lb_src_pkg_res.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}